chore: Move add tokens config into common widgets lib (#2488)
