+++ packages/checkout/widgets-lib/src/lib/squid/config.ts
@@ -1,3 +1,17 @@
 export const SQUID_SDK_BASE_URL = 'https://apiplus.squidrouter.com';
 
 export const SQUID_NATIVE_TOKEN = '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE';
+
+export const APPROVE_TXN_ANIMATION = '/access_coins.riv';
+
+export const EXECUTE_TXN_ANIMATION = '/swapping_coins.riv';
+
+export const BLOCK_TXN_ANIMATION = '/blocked.riv';
+
+export const ERROR_TXN_ANIMATION = '/error.riv';
+
+export const TOOLKIT_SQUID_URL = 'https://toolkit.immutable.com/squid-bridge/';
+
+export const FIXED_HANDOVER_DURATION = 2000;
+
+export const TOKEN_PRIORITY_ORDER = ['IMX', 'USDC', 'ETH'];
