--- packages/checkout/widgets-lib/src/widgets/add-tokens/utils/config.ts
@@ -1,13 +1,0 @@
-export const FIXED_HANDOVER_DURATION = 2000;
-
-export const APPROVE_TXN_ANIMATION = '/access_coins.riv';
-
-export const EXECUTE_TXN_ANIMATION = '/swapping_coins.riv';
-
-export const BLOCK_TXN_ANIMATION = '/blocked.riv';
-
-export const ERROR_TXN_ANIMATION = '/error.riv';
-
-export const TOKEN_PRIORITY_ORDER = ['IMX', 'USDC', 'ETH'];
-
-export const TOOLKIT_SQUID_URL = 'https://toolkit.immutable.com/squid-bridge/';
