+++ packages/checkout/widgets-lib/src/widgets/add-tokens/views/AddTokens.tsx
@@ -65,10 +65,9 @@ import { SquidFooter } from '../../../lib/squid/components/SquidFooter';
 import { TokenDrawerMenu } from '../components/TokenDrawerMenu';
 import { PULSE_SHADOW } from '../utils/animation';
 import { RouteData } from '../../../lib/squid/types';
-import { SQUID_NATIVE_TOKEN } from '../../../lib/squid/config';
+import { SQUID_NATIVE_TOKEN, TOOLKIT_SQUID_URL } from '../../../lib/squid/config';
 import { identifyUser } from '../../../lib/analytics/identifyUser';
 import { NotEnoughGasDrawer } from '../../../components/NotEnoughGasDrawer/NotEnoughGasDrawer';
-import { TOOLKIT_SQUID_URL } from '../utils/config';
 import { SelectedWallet } from '../../../components/SelectedWallet/SelectedWallet';
 import { RouteOptionsDrawer } from '../../../components/RouteOptionsDrawer/RouteOptionsDrawer';
 import { SelectedRouteOption } from '../../../components/SelectedRouteOption/SelectedRouteOption';
